+++ packages/core/lib/combo/patch-build/randomizer.ts
@@ -828,6 +828,7 @@ function worldConfig(world: World, settings: Settings): Set<Confvar> {
     MM_AUTO_INVERT_ALWAYS: settings.autoInvert === 'always',
     MM_AUTO_INVERT_FIRST_CYCLE: settings.autoInvert === 'firstCycle',
     MM_MOON_CRASH_CYCLE: settings.moonCrash === 'cycle',
+    OOT_OPEN_ZD_SHORTCUT: settings.openZdShortcut,
   };
 
   for (const v in exprs) {